requirements.txt
beautifulsoup4
dnspython
Requests
